requirements.txt
boto==2.44.0
python-dotenv==0.6.2
